+++ packages/babel-plugin-react-dom-render/src/index.test.ts
@@ -1,25 +1,42 @@
 import { transformSync } from '@babel/core'
+import { ComponentError } from '.'
 import plug from './index'
 const opts = {
   plugins: [plug],
   configFile: false,
-  compact: true,
+  compact: true, // All optional newlines and whitespace will be omitted
 }
+describe('validate component name', () => {
+  test('should throw with default class component', () => {
+    const code = 'export default class App extends React.Component {}'
+    expect(() => transformSync(code, opts)).toThrowError(
+      ComponentError.functionComponentOnly
+    )
+  })
+  test('should throw with anonymous function component', () => {
+    const code = 'export default function () {};'
+    expect(() => transformSync(code, opts)).toThrowError(
+      ComponentError.namedFunctionComponentOnly
+    )
+  })
+  test('should throw when naming function `VOID_JS_APP`', () => {
+    const code = 'export default function VOID_JS_APP() {};'
+    expect(() => transformSync(code, opts)).toThrowError(
+      ComponentError.reservedFunctionName
+    )
+  })
+})
 test('should import createElement and ReactDOM', () => {
-  const str = ''
-  const result = transformSync(str, opts)
+  const code = ''
+  const result = transformSync(code, opts)
   if (!result) return
   expect(result.code).toContain(
-    `import ReactDOM from"react-dom";import{createElement}from'react';`
+    `import ReactDOM from"react-dom";import{createElement}from"react";`
   )
 })
-test('should throw with default class component', () => {
-  const str = 'export default class App extends React.Component {}'
-  expect(() => transformSync(str, opts)).toThrowError()
-})
-test('should hydrate', () => {
-  const str = 'export default function App() {}'
-  const result = transformSync(str, {
+test('should use hydrate', () => {
+  const code = 'export default function App() {}'
+  const result = transformSync(code, {
     ...opts,
     plugins: [
       [
@@ -33,9 +50,43 @@ test('should hydrate', () => {
   if (!result) return
   expect(result.code).toContain('ReactDOM.hydrate')
 })
-test('should test existence of getStaticProps', () => {
-  const str = `export default function App() {}`
-  const result = transformSync(str, opts)
+test('should test against getStaticProps', () => {
+  const code = `export default function App() {}`
+  const result = transformSync(code, opts)
   if (!result) return
   expect(result.code).toContain('typeof getStaticProps')
+})
+describe('option `app`', () => {
+  it('should not import app', () => {
+    const code = `export default function App() {}`
+    const result = transformSync(code, {
+      ...opts,
+      plugins: [
+        [
+          plug,
+          {
+            app: false,
+          },
+        ],
+      ],
+    })
+    if (!result) return
+    expect(result.code).toContain('if(false)')
+  })
+  it('should import app', () => {
+    const code = `export default function App() {}`
+    const result = transformSync(code, {
+      ...opts,
+      plugins: [
+        [
+          plug,
+          {
+            app: '/path/to/app',
+          },
+        ],
+      ],
+    })
+    if (!result) return
+    expect(result.code).toContain('if(true)')
+  })
 })