+++ packages/doc/components/Footer.js
@@ -1,17 +1,15 @@
 export default function Footer() {
   return (
-    <footer className="text-center mt-32 text-gray-900">
-      Built with <strong className="font-bold text-black">voidjs</strong>
-      <div>
-        <a
-          target="_blank"
-          href="https://github.com/chenxsan/voidjs/tree/master/packages/doc"
-          rel="noopener noreferrer"
-          className="underline text-blue-700 hover:text-blue-900 transition duration-200"
-        >
-          Source code
-        </a>
-      </div>
+    <footer className="mt-32 text-gray-900 text-md">
+      Built with{' '}
+      <a
+        target="_blank"
+        href="https://github.com/chenxsan/voidjs/tree/master/packages/doc"
+        rel="noopener noreferrer"
+        className="font-normal text-blue-700 hover:text-blue-900 transition duration-200"
+      >
+        voidjs
+      </a>
     </footer>
   )
 }