+++ packages/doc/pages/index.js
@@ -1,24 +1,17 @@
-// polyfill for ie 11
-import 'react-app-polyfill/ie11'
-
-import 'prismjs/themes/prism-okaidia.css'
-import 'css/tailwind.css'
-import '@void-js/styles/lib/index.css'
 import logo from 'img/voidjs-logo.svg'
 import favicon from 'img/favicon.png'
-import Footer from '../components/Footer'
 import Content from '../markdown/home.md'
 import { Helmet } from 'react-helmet'
 export default function Home(props) {
   return (
-    <div className="container mx-auto">
+    <div className="container mx-auto px-5 lg:px-0 ">
       <Helmet>
         <html lang="en" />
         <title>{props.title}</title>
         <meta name="description" content={props.description} />
         <link rel="icon" type="image/png" href={favicon} />
       </Helmet>
-      <h1 className="flex text-4xl leading-none justify-center pt-6 lg:text-4xl font-bold items-center">
+      <h1 className="flex text-4xl leading-none justify-start pt-6 lg:text-4xl font-bold font-serif">
         <img
           src={logo}
           alt="voidjs logo"
@@ -26,13 +19,12 @@ export default function Home(props) {
         />
         {props.title}
       </h1>
-      <h2 className="flex justify-center text-lg pt-0 px-2 leading-tight font-normal">
+      <h2 className="flex justify-start text-sm pt-0 leading-tight font-normal my-2">
         {props.description}
       </h2>
       <div className="voidjs">
         <Content />
       </div>
-      <Footer />
     </div>
   )
 }