+++ packages/voidjs-cli/src/DevServer/createWebpackConfig.ts
@@ -1,29 +1,34 @@
 import webpack from 'webpack'
-import { extensions, alias, postcssPlugins } from '../config'
+import {
+  resolveExtensions,
+  alias,
+  postcssPlugins,
+  supportedImageExtensions,
+  supportedFontExtensions,
+  supportedCssExtensions,
+} from '../config'
 import rehypePrism from '@mapbox/rehype-prism'
 import createEntries from './createEntries'
-import type { EntryObject } from './index'
+import path from 'path'
+import isPageEntry from '../utils/isPageEntry'
 
-const createDOMRenderRule = (pagesDir: string) => ({
-  include: (filename: string): boolean => {
-    // entries under pagesDir
-    // exclude client entry
-    return (
-      filename.startsWith(pagesDir) && filename.includes('.client.') === false
-    )
-  },
+import type { RuleSetRule, Configuration } from 'webpack'
+
+const createDOMRenderRule = (pagesDir: string, hasCustomApp: boolean) => ({
+  include: (filename: string): boolean => isPageEntry(pagesDir, filename),
   plugins: [
     [
-      'react-dom-render', // render page entry in dom
+      'react-dom-render',
       {
         hydrate: false,
         root: 'voidjs-app',
+        app: hasCustomApp === true ? path.join(pagesDir, '_app') : hasCustomApp,
       },
     ],
   ],
 })
 
-const createBabelOptions = (pagesDir: string) => {
+const createBabelOptions = (pagesDir: string, hasCustomApp: boolean) => {
   return {
     presets: [
       '@babel/preset-env',
@@ -48,19 +53,34 @@ const createBabelOptions = (pagesDir: string) => {
         },
       ],
     ],
-    overrides: [createDOMRenderRule(pagesDir)],
+    overrides: [createDOMRenderRule(pagesDir, hasCustomApp)],
   }
 }
-
 export default function createWebpackConfig(
   activePages: string[],
   pagesDir: string,
+  hasCustomApp: boolean,
   socketUrl: string
-): webpack.Configuration {
+): Configuration {
+  const babelLoader: RuleSetRule = {
+    loader: 'babel-loader',
+    options: createBabelOptions(pagesDir, hasCustomApp),
+  }
+
   return {
     experiments: {
       topLevelAwait: true,
     },
+    // TODO disable at the moment as I'm not yet wrap my head around it
+    // cache: {
+    //   type: 'filesystem',
+    //   buildDependencies: {
+    //     // https://github.com/webpack/changelog-v5/blob/master/guides/persistent-caching.md
+    //     pages: [path.resolve(pagesDir, path.sep)],
+    //     // FIXME might have bugs
+    //   },
+    // },
+    devtool: 'eval',
     mode: 'development',
     // TODO should be configurable
     // because people might not care ie 11 like me
@@ -68,7 +88,7 @@ export default function createWebpackConfig(
     // we use function for entry
     // so it would be called every compilation
     // see https://github.com/webpack/webpack/pull/10734#issuecomment-616198739
-    entry: (): EntryObject => createEntries(pagesDir, activePages),
+    entry: () => createEntries(pagesDir, activePages),
     output: {
       publicPath: '/',
     },
@@ -79,46 +99,38 @@ export default function createWebpackConfig(
         {
           test: /\.(mjs|js|jsx|ts|tsx)$/i,
           exclude: [/node_modules/],
-          use: [
-            {
-              loader: 'babel-loader',
-              options: createBabelOptions(pagesDir),
-            },
-          ],
+          use: [babelLoader],
         },
         {
           test: /\.(mdx|md)$/,
           use: [
-            {
-              loader: 'babel-loader',
-              options: createBabelOptions(pagesDir),
-            },
+            babelLoader,
             {
               loader: '@mdx-js/loader',
               options: {
-                rehypePlugins: [rehypePrism],
+                rehypePlugins: [rehypePrism], // support code highlighting
               },
             },
           ],
         },
         {
-          test: /\.(png|svg|jpg|jpeg|gif)$/i,
+          test: supportedImageExtensions,
           type: 'asset',
         },
         {
-          test: /\.(woff(2)?|ttf|eot)(\?v=\d+\.\d+\.\d+)?$/,
+          test: supportedFontExtensions,
           type: 'asset/resource',
           generator: {
             filename: 'fonts/[name][ext]',
           },
         },
         {
-          test: /\.(sa|sc|c)ss$/i,
+          test: supportedCssExtensions,
           use: [
             'style-loader',
             'css-loader',
             {
-              loader: 'postcss-loader',
+              loader: 'postcss-loader', // FIXME this is slow, might be related to tailwindcss https://github.com/tailwindlabs/tailwindcss/issues/2544#issuecomment-710974396
               options: {
                 postcssOptions: {
                   ident: 'postcss',
@@ -126,14 +138,14 @@ export default function createWebpackConfig(
                 },
               },
             },
-            'sass-loader',
+            'sass-loader', // built-in sass support
           ],
         },
       ],
     },
     resolve: {
-      extensions,
-      alias,
+      extensions: resolveExtensions,
+      alias, // TODO we should let users configure this
     },
     plugins: [
       new webpack.DefinePlugin({
@@ -141,6 +153,9 @@ export default function createWebpackConfig(
         __WEBSOCKET__: JSON.stringify(socketUrl),
       }),
       new webpack.NoEmitOnErrorsPlugin(),
+      new webpack.ProgressPlugin({
+        profile: true, // enable profiling
+      }),
     ],
   }
 }