+++ packages/voidjs-cli/src/DevServer/index.ts
@@ -1,8 +1,8 @@
 /**
  * Copyright 2020-present, Sam Chen.
- * 
+ *
  * Licensed under GPL-3.0-or-later
- * 
+ *
  * This file is part of voidjs.
 
     voidjs is free software: you can redistribute it and/or modify
@@ -20,7 +20,7 @@
  */
 import webpack from 'webpack'
 import * as path from 'path'
-import { logger, publicFolder } from '../config'
+import { logger, publicFolder, socketPath } from '../config'
 import express from 'express'
 import webpackDevMiddleware from 'webpack-dev-middleware'
 import http from 'http'
@@ -31,7 +31,9 @@ import watchCompilation from './watchCompilation'
 import createWebSocketServer from './createWebSocketServer'
 import Builder from '../Builder'
 import HtmlPlugin from '../webpack-plugins/HtmlPluginForDevServer'
+import hasCustomApp from '../utils/hasCustomApp'
 
+// TODO to be removed after webpack export EntryObject typing
 export interface EntryObject {
   [index: string]:
     | [string, ...string[]]
@@ -48,9 +50,7 @@ export interface Server {
 export default class DevServer extends Builder {
   readonly #host: string
   readonly #port: number
-
-  // collect activated pages
-  #activePages: string[]
+  #activePages: string[] // collect activated pages
 
   constructor(
     pagesDir: string,
@@ -60,24 +60,21 @@ export default class DevServer extends Builder {
 
     this.#host = host
     this.#port = port
-
     this.#activePages = []
   }
 
   public async start(): Promise<http.Server> {
-    // resolve voidjs.config.js
-    await this.resolveConfig()
-
-    const socketPath = '/__websocket'
-
     const webpackConfig = createWebpackConfig(
       this.#activePages,
       this.pagesDir,
+      hasCustomApp(this.pagesDir),
       `${this.#host}:${this.#port}${socketPath}`
     )
-    const compiler = webpack(webpackConfig)
 
+    const compiler = webpack(webpackConfig)
     const devMiddleware = webpackDevMiddleware(compiler)
+
+    // append entry only when accessed
     const voidjsMiddleware = (pagesDir: string) => (
       req: express.Request,
       _res: express.Response,
@@ -88,12 +85,11 @@ export default class DevServer extends Builder {
         // check if page does exit on disk
         // FIXME what if req.url contains query
         const page = findSourceFile(pagesDir, req.url)
-
-        if (page.exists) {
-          const src = page.src as string
-          if (!this.#activePages.includes(src)) {
-            this.#activePages.push(src)
-            new HtmlPlugin(pagesDir, src).apply(compiler)
+        if (page.exists === true) {
+          const pageSrc = page.src as string
+          if (!this.#activePages.includes(pageSrc)) {
+            this.#activePages.push(pageSrc)
+            new HtmlPlugin(pagesDir, pageSrc).apply(compiler)
             devMiddleware.invalidate()
           }
         }
@@ -104,6 +100,7 @@ export default class DevServer extends Builder {
     const app = express()
     /**
      * rewrite html request
+     * TODO tests needed
      */
     app.use(function (
       req: express.Request,
@@ -121,6 +118,7 @@ export default class DevServer extends Builder {
       }
       next()
     })
+
     app.use(voidjsMiddleware(this.pagesDir))
     app.use(devMiddleware)